Not because you can it means your should.
And with those words in mind...
So I got tired that every time I needed to update the settings I had to
navigate trough hundreds of lines to find the one I'm looking for, so I
diced to abuse the import functionality of python to separate the
settings file into multiple files.
I tried to name each file name self explanatory on what contains. And
also third party settings are on his own module, in this case each file
has the name of the library that belongs to.
I have no idea if this is a good idea or a bad one, but is something I
want to help me navigate the setting better, and I think I got a good
result.